added Window class

+++ Application/Animations/Animation.cpp
@@ -34,12 +34,13 @@ namespace Engine
     Animation& Animation::add_frames_line(int number_x,int number_y,int line)
     {
         const sf::Vector2u size = m_texture->getSize();
-        const float delta_x = size.x / float(number_x);
-        const float delta_y = size.y / float(number_y);
+        const int delta_x = size.x / number_x;
+        const int delta_y = size.y / number_y;
 
         for(int i = 0;i<number_x;++i)
         {
-            add_frame(sf::IntRect({i*delta_x,line*delta_y},{delta_x,delta_y}));
+            add_frame(sf::IntRect({i*delta_x,line*delta_y},
+                {delta_x,delta_y}));
         }
 
         return *this;
@@ -48,8 +49,8 @@ namespace Engine
     Animation& Animation::add_frames_column(int number_x,int number_y,int column)
     {
         const sf::Vector2u size = m_texture->getSize();
-        const float delta_x = size.x / float(number_x);
-        const float delta_y = size.y / float(number_y);
+        const int delta_x = size.x / number_x;
+        const int delta_y = size.y / number_y;
         
         for(int i = 0;i<number_y;++i)
         {
